requirements.txt
alpaca-py>=0.28.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
click>=8.1.0
tabulate>=0.9.0
